requirements.txt
fastapi==0.95.1
#gradio==4.16.0
#gradio_imageslider==0.0.17
#gradio_client==0.8.1
Markdown==3.4.1
numpy==1.24.2
requests==2.28.2
sentencepiece==0.1.98
tokenizers==0.13.3
#torch>=2.1.0
#torchvision>=0.16.0
uvicorn==0.21.1
wandb==0.14.0
httpx==0.24.0
transformers==4.28.1
#accelerate==0.18.0 # torch-conflict
scikit-learn==1.2.2
sentencepiece==0.1.98
einops==0.7.0
einops-exts==0.0.4
#timm==0.9.8 # torch-conflict
openai-clip==1.0.1
fsspec==2023.4.0
#kornia==0.6.9 # torch-conflict
matplotlib==3.7.1
ninja==1.11.1
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.7.0.72
pandas==2.0.1
Pillow==9.4.0
pytorch-lightning==2.4.0
PyYAML==6.0
scipy==1.9.1
tqdm==4.65.0
triton==2.1.0
urllib3==1.26.15
webdataset==0.2.48
#xformers==0.0.22.post4+cu118 # torch-conflict
#facexlib==0.3.0 # torch-conflict
#k-diffusion==0.1.1.post1 # torch-conflict
diffusers==0.16.1
